move walkings trajectoryUtils into walking
before it was located in bitbots_splines but since it is exclusively used and adapted to bitbots_quintic_walk it is now moved there
